+++ src/webview/modelPromptStructures.ts
@@ -1,7 +1,0 @@
-/* eslint-disable @typescript-eslint/naming-convention */
-export const claude_2_promptStructure = `Human: 
-
-Assistant:
-`;
-
-export const notClaudeCommandPromptStructure = ``;